#8: Mov waypoint related views to their own components
