+++ src/app/pages/create-trip-page/components/city-list-panel/city-list-panel.component.html
@@ -6,206 +6,47 @@
     <!-- Cities List View -->
 
     @case (PanelView.CitiesListView) {
-      <app-city-list
+    <app-city-list
       [setCurrentView]="setCurrentView.bind(this)"
       [openDeleteModal]="openDeleteModal.bind(this)"
-      style="width: 100%;">
-      </app-city-list>
+      style="width: 100%"
+    >
+    </app-city-list>
     }
 
     <!-- City View -->
 
     @case (PanelView.CityView) {
     <app-city-form
-    [onDiscardClick]="onDiscardClick.bind(this)"
-    [setCurrentView]="setCurrentView.bind(this)"
-    [preventEnterSubmit]="preventEnterSubmit.bind(this)"
-    style="width: 100%;">
-    >  
+      [onDiscardClick]="onDiscardClick.bind(this)"
+      [setCurrentView]="setCurrentView.bind(this)"
+      [preventEnterSubmit]="preventEnterSubmit.bind(this)"
+      style="width: 100%"
+    >
+      >
     </app-city-form>
     }
 
     <!-- Waypoints List View -->
 
     @case (PanelView.WaypointsListView) {
-    <div class="day-navigation">
-      <button
-        [class.hidden]="currentDayIndex === 0"
-        (click)="navigateToDay(currentDayIndex - 1)"
-        class="arrow"
-      >
-        &#9664;
-      </button>
-      <h3 class="day-header">
-        {{ getDateForDay(selectedCity!.arrivalDate!, currentDayIndex) }}
-      </h3>
-      <button
-        [class.hidden]="
-          currentDayIndex === (selectedCity?.waypoints)!.length - 1
-        "
-        (click)="navigateToDay(currentDayIndex + 1)"
-        class="arrow"
-      >
-        &#9654;
-      </button>
-    </div>
-    <div
-      *ngFor="let waypoint of selectedCity!.waypoints[currentDayIndex]"
-      class="option-button"
+    <app-waypoint-list
+      [setCurrentView]="setCurrentView.bind(this)"
+      [openDeleteModal]="openDeleteModal.bind(this)"
+      style="width: 100%"
     >
-      <div class="button-info-container">
-        <span class="button-text-container">
-          <div class="button-icon" id="city-icon"></div>
-          <span class="button-text name">{{ waypoint.name }}</span>
-        </span>
-        <div class="date-text">
-          {{ waypoint.startTime }} - {{ waypoint.endTime }}
-        </div>
-      </div>
-      <div class="menu-container" (click)="$event.stopPropagation()">
-        <div class="menu-dots">•••</div>
-        <div class="dropdown-menu">
-          <div
-            class="dropdown-item"
-            id="edit"
-            (click)="onEditClick(waypoint, 'waypoint')"
-          >
-            Edit
-          </div>
-          <div
-            class="dropdown-item"
-            id="delete"
-            (click)="onDeleteClick(waypoint, 'waypoint')"
-          >
-            Delete
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="option-button" (click)="setCurrentView(PanelView.WaypointView)">
-      <span class="button-text-container">
-        <div class="button-icon" id="add-icon"></div>
-        <span class="button-text">Add new waypoint</span>
-      </span>
-    </div>
+    </app-waypoint-list>
     }
 
     <!-- Waypoint View -->
 
     @case (PanelView.WaypointView) {
-    <form
-      class="add-form"
-      [formGroup]="waypointForm"
-      (ngSubmit)="onWaypointFormSubmit()"
-      (keydown.enter)="preventEnterSubmit($event)"
-    >
-      <div class="form-row-section">
-        <label class="form-label">Waypoint Name</label>
-        <input
-          #waypointName
-          type="text"
-          formControlName="waypointName"
-          placeholder="Enter the waypoint name"
-          [class.disabled-input]="waypointForm.get('waypointName')?.disabled"
-        />
-        <small
-          *ngIf="
-            waypointForm.get('waypointName')?.hasError('required') &&
-            waypointForm.get('waypointName')?.touched
-          "
-          class="text-danger"
-        >
-          Waypoint Name cannot be empty.
-        </small>
-        <small
-          *ngIf="waypointForm.get('waypointName')?.hasError('noResult')"
-          class="text-danger"
-        >
-          No results found. Please select a valid waypoint from the list.
-        </small>
-        <small
-          *ngIf="
-            waypointForm.get('waypointName')?.hasError('unsupportedType') &&
-            waypointForm.get('waypointName')?.touched
-          "
-          class="text-danger"
-        >
-          The waypoint selected is not a food location, an attraction or a
-          recreational place.
-        </small>
-      </div>
-      <div class="form-row-section number">
-        <label class="form-label number">Start Time</label>
-        <span class="time-counter">
-          <input
-            class="number"
-            type="number"
-            formControlName="startHour"
-            min="0"
-            max="23"
-            (input)="formatTwoDigitsInput($event, 'hour')"
-          />
-          h
-          <input
-            class="number"
-            type="number"
-            formControlName="startMinutes"
-            min="0"
-            max="59"
-            (input)="formatTwoDigitsInput($event, 'minutes')"
-          />
-          m
-        </span>
-      </div>
-      <div class="form-row-section number">
-        <label class="form-label number">End Time</label>
-        <span class="time-counter">
-          <input
-            class="number"
-            type="number"
-            formControlName="endHour"
-            min="0"
-            max="23"
-            (input)="formatTwoDigitsInput($event, 'hour')"
-          />
-          h
-          <input
-            class="number"
-            type="number"
-            formControlName="endMinutes"
-            min="0"
-            max="59"
-            (input)="formatTwoDigitsInput($event, 'minutes')"
-          />
-          m
-        </span>
-      </div>
-      <div class="form-row-section">
-        <small
-          *ngIf="waypointForm.hasError('endTimeEarlier')"
-          class="text-danger"
-        >
-          End time must be later than the start time.
-        </small>
-        <small
-          *ngIf="waypointForm.hasError('overlap')"
-          class="text-danger"
-        >
-          Current interval is overlapping with one of the existing ones.
-        </small>
-      </div>
-      <div class="buttons-section">
-        <div
-          class="discard-button"
-          (click)="onDiscardClick(PanelView.WaypointsListView, 'waypoint')"
-        >
-          Discard
-        </div>
-        <button type="submit" class="submit-button">
-          {{ isEditFlow === false ? "Add" : "Edit" }}
-        </button>
-      </div>
-    </form>
+    <app-waypoint-form
+      [onDiscardClick]="onDiscardClick.bind(this)"
+      [setCurrentView]="setCurrentView.bind(this)"
+      [preventEnterSubmit]="preventEnterSubmit.bind(this)"
+      style="width: 100%"
+    ></app-waypoint-form>
     } }
   </div>
 </div>
