+++ src/app/pages/create-trip-page/components/city-list-panel/city-list-panel.component.ts
@@ -45,27 +45,7 @@ export class CityListPanelComponent
   @Input() viewChanged: PanelView | null = null;
   @Input() openDeleteModal!: (type: 'city' | 'waypoint', data: any) => void;
 
-  @Output() action = new EventEmitter<{
-    type:
-      | 'citySubmitted'
-      | 'cityEdited'
-      | 'waypointSubmitted'
-      | 'waypointEdited'
-      | 'citySelected'
-      | 'dayChanged';
-    payload?: any;
-  }>();
-
-  @ViewChild('waypointName') waypointNameInput?: ElementRef<HTMLInputElement>;
-
   PanelView = PanelView;
-  private waypointAutocomplete: google.maps.places.Autocomplete | null = null;
-  private getWaypoints = () =>
-    this.selectedCity?.waypoints[this.currentDayIndex] ?? undefined;
-  private getIsEditFlag = () => ({
-    isEditFlow: this.isEditFlow ?? false,
-    waypointInProcess: this.waypointInProcess ?? undefined,
-  });
 
   waypointForm: FormGroup = new FormGroup({});
 
@@ -78,7 +58,6 @@ export class CityListPanelComponent
   private subscriptions: Subscription[] = [];
 
   constructor(
-    private formBuilder: FormBuilder,
     private changeDetector: ChangeDetectorRef,
     private googleMapsService: GoogleMapsService,
     private tripStateService: TripStateService
@@ -93,25 +72,12 @@ export class CityListPanelComponent
         this.handleCitySelected(selectedCity);
       }),
       this.tripStateService.getCityToAdd().subscribe((cityToAdd) => {
-        if(cityToAdd === undefined){
+        if (cityToAdd === undefined) {
           return;
         }
 
         this.setCurrentView(PanelView.CityView);
       })
-    );
-
-    this.waypointForm = this.formBuilder.group(
-      {
-        waypointName: ['', [Validators.required]],
-        startHour: ['00', [Validators.required]],
-        startMinutes: ['00', [Validators.required]],
-        endHour: ['00', [Validators.required]],
-        endMinutes: ['01', [Validators.required]],
-      },
-      {
-        validators: timeValidator(this.getWaypoints, this.getIsEditFlag),
-      }
     );
 
     try {
@@ -178,87 +144,6 @@ export class CityListPanelComponent
     this.changeDetector.detectChanges();
   }
 
-  onWaypointFormSubmit() {
-    if (!this.waypointForm.valid) {
-      console.log('Invalid Form: ' + this.waypointForm.errors);
-      return;
-    }
-
-    if (this.waypointInProcess === null) {
-      this.waypointForm.get('waypointName')?.setErrors({ noResult: true });
-      return;
-    }
-
-    const startHour = this.waypointForm.get('startHour')?.value;
-    const startMinutes = this.waypointForm.get('startMinutes')?.value;
-
-    const endHour = this.waypointForm.get('endHour')?.value;
-    const endMinutes = this.waypointForm.get('endMinutes')?.value;
-
-    this.waypointInProcess!.startTime = `${this.formatTwoDigits(
-      startHour,
-      'hour'
-    )}:${this.formatTwoDigits(startMinutes, 'minutes')}`;
-    this.waypointInProcess!.endTime = `${this.formatTwoDigits(
-      endHour,
-      'hour'
-    )}:${this.formatTwoDigits(endMinutes, 'minutes')}`;
-
-    let response: boolean;
-    if (this.isEditFlow) {
-      response = this.HandleWaypointEdit();
-      this.isEditFlow = false;
-    } else {
-      response = this.HandleWaypointAdd();
-    }
-
-    if (response) {
-      this.waypointForm.get('waypointName')?.enable();
-      this.setCurrentView(PanelView.WaypointsListView);
-      this.waypointInProcess = null;
-    }
-  }
-
-  onCityClick(selectedCity: AddCityDto) {
-    console.log('City clicked: ' + selectedCity.name);
-    const cityBounds = new google.maps.LatLngBounds(
-      new google.maps.LatLng(
-        selectedCity.southWestBound.latitude,
-        selectedCity.southWestBound.longitude
-      ),
-      new google.maps.LatLng(
-        selectedCity.northEastBound.latitude,
-        selectedCity.northEastBound.longitude
-      )
-    );
-
-    this.selectedCity = selectedCity;
-    this.setCurrentView(PanelView.WaypointsListView);
-
-    this.action.emit({
-      type: 'citySelected',
-      payload: new SelectedCityDto(selectedCity, cityBounds),
-    });
-  }
-
-  onEditClick(entity: AddCityDto | AddWaypointDto, type: 'city' | 'waypoint') {
-    this.isEditFlow = true;
-
-    if (type === 'waypoint') {
-      this.setCurrentView(PanelView.WaypointView);
-      this.setWaypointForm(entity as AddWaypointDto);
-    }
-
-    this.changeDetector.detectChanges();
-  }
-
-  onDeleteClick(
-    entity: AddCityDto | AddWaypointDto,
-    type: 'city' | 'waypoint'
-  ) {
-    this.openDeleteModal(type, entity);
-  }
-
   onDiscardClick(view: PanelView, type: 'city' | 'waypoint') {
     if (this.isEditFlow) {
       this.isEditFlow = false;
@@ -276,54 +161,8 @@ export class CityListPanelComponent
   }
 
   setCurrentView(view: PanelView) {
-    if (this.currentView === PanelView.WaypointView) {
-      this.destroyAutocomplete(this.waypointAutocomplete);
-    }
-
     this.currentView = view;
     this.changeDetector.detectChanges();
-
-    switch (view) {
-      case PanelView.CitiesListView:
-        break;
-      case PanelView.WaypointView:
-        this.waypointForm.reset({
-          startHour: '00',
-          startMinutes: '00',
-          endHour: '00',
-          endMinutes: '01',
-        });
-        this.initializeWaypointAutocomplete();
-        break;
-      case PanelView.WaypointsListView:
-        break;
-      default:
-        break;
-    }
-  }
-
-  navigateToDay(dayIndex: number): void {
-    this.currentDayIndex = dayIndex;
-
-    this.action.emit({
-      type: 'dayChanged',
-      payload: { dayIndex: this.currentDayIndex },
-    });
-  }
-
-  getDateForDay(startDate: Date, dayIndex: number): string {
-    const result = new Date(startDate);
-    result.setDate(startDate.getDate() + dayIndex);
-    return result.toLocaleDateString('en-US', {
-      day: 'numeric',
-      month: 'long',
-    });
-  }
-
-  formatTwoDigitsInput(event: Event, type: 'hour' | 'minutes'): void {
-    const inputElement = event.target as HTMLInputElement;
-
-    inputElement.value = this.formatTwoDigits(inputElement.value, type);
   }
 
   preventEnterSubmit(event: Event) {
@@ -352,154 +191,5 @@ export class CityListPanelComponent
     if (index > -1 && index < focusableElements.length - 1) {
       focusableElements[index + 1].focus();
     }
-  }
-
-  private HandleWaypointAdd(): boolean {
-    const enteredValue = this.waypointNameInput?.nativeElement.value.trim();
-    if (!enteredValue || this.waypointInProcess!.name !== enteredValue) {
-      this.waypointForm.get('waypointName')?.setErrors({ noResult: true });
-      return false;
-    }
-
-    this.action.emit({
-      type: 'waypointSubmitted',
-      payload: this.waypointInProcess,
-    });
-    return true;
-  }
-
-  private HandleWaypointEdit(): boolean {
-    this.action.emit({
-      type: 'waypointEdited',
-      payload: this.waypointInProcess,
-    });
-    return true;
-  }
-
-  private setWaypointForm(waypointToEdit: AddWaypointDto) {
-    this.waypointInProcess = waypointToEdit;
-
-    this.waypointForm.get('waypointName')?.disable();
-    const [startHour, startMinutes] = waypointToEdit.startTime.split(':');
-    const [endHour, endMinutes] = waypointToEdit.endTime.split(':');
-    this.waypointForm.patchValue({
-      waypointName: waypointToEdit.name,
-      startHour: startHour,
-      startMinutes: startMinutes,
-      endHour: endHour,
-      endMinutes: endMinutes,
-    });
-  }
-
-  private initializeWaypointAutocomplete() {
-    if (this.waypointNameInput?.nativeElement === undefined) {
-      console.error(
-        'Waypoint Name Input is not rendered. The autocomplete cannot be initialized.'
-      );
-      return;
-    }
-
-    if (this.selectedCity === null) {
-      console.error(
-        'No city is selected. The autocomplete cannot be initialized.'
-      );
-      return;
-    }
-
-    this.waypointAutocomplete = new google.maps.places.Autocomplete(
-      this.waypointNameInput!.nativeElement,
-      {
-        types: ['establishment'],
-        fields: ['name', 'types', 'place_id', 'geometry'],
-        bounds: new google.maps.LatLngBounds(
-          new google.maps.LatLng(
-            this.selectedCity.southWestBound.latitude,
-            this.selectedCity.southWestBound.longitude
-          ),
-          new google.maps.LatLng(
-            this.selectedCity.northEastBound.latitude,
-            this.selectedCity.northEastBound.longitude
-          )
-        ),
-        strictBounds: true,
-      }
-    );
-
-    this.waypointAutocomplete.addListener('place_changed', () => {
-      let place = this.waypointAutocomplete!.getPlace();
-
-      const waypointName = place.name?.split(',')[0] ?? '';
-      this.waypointNameInput!.nativeElement.value = waypointName;
-
-      if (!place.place_id || !place.name) {
-        return;
-      }
-
-      const waypointType = this.getWaypointType(place.types!);
-      if (waypointType === WaypointType.Unkwnown) {
-        this.waypointForm.get('waypointName')?.setErrors({
-          unsupportedType: true,
-        });
-        return;
-      }
-
-      // this.waypointForm.get('waypointName')?.setErrors(null);
-
-      // IMPORTANT: Show no result feedback, country filtering etc
-      const placeId = place.place_id ?? '';
-      const latitude = place.geometry?.location?.lat() ?? 0;
-      const longitude = place.geometry?.location?.lng() ?? 0;
-
-      this.waypointInProcess = new AddWaypointDto(
-        waypointName,
-        waypointType.toString(),
-        placeId,
-        latitude,
-        longitude,
-        '',
-        '',
-        0
-      );
-    });
-  }
-
-  private destroyAutocomplete(
-    autocomplete: google.maps.places.Autocomplete | null
-  ) {
-    if (autocomplete) {
-      google.maps.event.clearInstanceListeners(autocomplete);
-      autocomplete = null;
-    }
-  }
-
-  private getWaypointType(types: string[]): WaypointType {
-    const foodMatch = types.find((type) =>
-      FOOD_WAYPOINT_TYPES.find((knownType) => knownType === type)
-    );
-    if (foodMatch) {
-      return WaypointType.Food;
-    }
-
-    const recreationalMatch = types.find((type) =>
-      RECREATIONAL_WAYPOINT_TYPES.find((knownType) => knownType === type)
-    );
-    if (recreationalMatch) {
-      return WaypointType.Recreational;
-    }
-
-    const attractionMatch = types.find((type) =>
-      ATTRACTIONS_WAYPOINT_TYPES.find((knownType) => knownType === type)
-    );
-    if (attractionMatch) {
-      return WaypointType.Attraction;
-    }
-
-    return WaypointType.Unkwnown;
-  }
-
-  private formatTwoDigits(text: string, type: 'hour' | 'minutes'): string {
-    let value = parseInt(text, 10);
-
-    return value.toString().padStart(2, '0');
   }
 }
