+++ src/app/pages/create-trip-page/components/city-list-panel/components/waypoint-form/waypoint-form.component.scss
@@ -1,0 +1,1 @@
+@import "../../city-list-panel.component.scss";
