+++ src/app/pages/create-trip-page/components/city-list-panel/components/waypoint-list/waypoint-list.component.html
@@ -1,0 +1,60 @@
+<div class="day-navigation">
+    <button
+      [class.hidden]="currentDayIndex === 0"
+      (click)="navigateToDay(currentDayIndex - 1)"
+      class="arrow"
+    >
+      &#9664;
+    </button>
+    <h3 class="day-header">
+      {{ getDateForDay(selectedCity!.arrivalDate!, currentDayIndex) }}
+    </h3>
+    <button
+      [class.hidden]="
+        currentDayIndex === (selectedCity?.waypoints)!.length - 1
+      "
+      (click)="navigateToDay(currentDayIndex + 1)"
+      class="arrow"
+    >
+      &#9654;
+    </button>
+  </div>
+  <div
+    *ngFor="let waypoint of selectedCity!.waypoints[currentDayIndex]"
+    class="option-button"
+  >
+    <div class="button-info-container">
+      <span class="button-text-container">
+        <div class="button-icon" id="city-icon"></div>
+        <span class="button-text name">{{ waypoint.name }}</span>
+      </span>
+      <div class="date-text">
+        {{ waypoint.startTime }} - {{ waypoint.endTime }}
+      </div>
+    </div>
+    <div class="menu-container" (click)="$event.stopPropagation()">
+      <div class="menu-dots">•••</div>
+      <div class="dropdown-menu">
+        <div
+          class="dropdown-item"
+          id="edit"
+          (click)="onEditClick(waypoint)"
+        >
+          Edit
+        </div>
+        <div
+          class="dropdown-item"
+          id="delete"
+          (click)="onDeleteClick(waypoint)"
+        >
+          Delete
+        </div>
+      </div>
+    </div>
+  </div>
+  <div class="option-button" (click)="setCurrentView(PanelView.WaypointView)">
+    <span class="button-text-container">
+      <div class="button-icon" id="add-icon"></div>
+      <span class="button-text">Add new waypoint</span>
+    </span>
+  </div>
