+++ src/app/pages/create-trip-page/create-trip-page.component.ts
@@ -51,7 +51,6 @@ export class CreateTripPageComponent implements OnInit, OnDestroy {
   private subscriptions: Subscription[] = [];
 
   constructor(
-    private googleMapsService: GoogleMapsService,
     private changeDetector: ChangeDetectorRef,
     private tripState: TripStateService
   ) {}
@@ -70,31 +69,6 @@ export class CreateTripPageComponent implements OnInit, OnDestroy {
     this.subscriptions.forEach(sub => sub.unsubscribe());
   }
 
-  handlePanelAction(event: { type: string; payload?: any }) {
-    switch (event.type) {
-      case 'citySubmitted':
-        this.onCitySubmitted(event.payload);
-        break;
-      case 'cityEdited':
-        this.onCityEdited(event.payload);
-        break;
-      case 'waypointSubmitted':
-        this.onWaypointSubmitted(event.payload);
-        break;
-      case 'waypointEdited':
-        this.onWaypointEdited(event.payload);
-        break;
-      case 'citySelected':
-        this.onCitySelected(event.payload);
-        break;
-      case 'dayChanged':
-        this.onDayChanged(event.payload.dayIndex);
-        break;
-      default:
-        console.error('Unhandled event type:', event.type);
-    }
-  }
-
   onCityToAdd(cityToAddData: { city: CityTransferDto }): void {
     this.cityToAdd = cityToAddData.city;
     this.changeDetector.detectChanges();
@@ -105,69 +79,6 @@ export class CreateTripPageComponent implements OnInit, OnDestroy {
     this.startDate = tripStarted.startDate;
 
     this.changeDetector.detectChanges();
-  }
-
-  onCitySubmitted(city: AddCityDto): void {
-    const updatedCities = [...this.cities, city];
-    this.tripState.updateCities(updatedCities);
-  }
-
-  onCityEdited(cityList: AddCityDto[]): void {
-    this.cities = [...cityList];
-
-    this.changeDetector.detectChanges();
-  }
-
-  onWaypointSubmitted(waypoint: AddWaypointDto): void {
-    console.log(this.cities);
-
-    let city = this.cities.find(
-      (city) => city === this.selectedCity?.selectedCity
-    );
-
-    this.setWaypointsOrder(
-      waypoint,
-      this.selectedCity!.selectedCity!.waypoints[this.currentDayIndex]
-        .length,
-      city!
-    );
-
-    city?.waypoints[this.currentDayIndex].push(waypoint);
-    city?.waypoints[this.currentDayIndex].sort((a, b) => a.order - b.order);
-    this.cities = [...this.cities];
-    console.log(this.cities);
-
-    this.changeDetector.detectChanges();
-  }
-
-  onWaypointEdited(waypoint: AddWaypointDto): void {
-    let city = this.cities.find(
-      (city) => city === this.selectedCity?.selectedCity
-    );
-
-    let waypointInList = city?.waypoints[this.currentDayIndex].find(
-      (cityWaypoint) => cityWaypoint.placeId === waypoint.placeId
-    );
-
-    waypointInList!.startTime = waypoint.startTime;
-    waypointInList!.endTime = waypoint.endTime;
-
-    this.setWaypointsOrder(waypointInList!, waypointInList!.order, city!);
-
-    city?.waypoints[this.currentDayIndex].sort((a, b) => a.order - b.order);
-    this.cities = [...this.cities];
-    console.log(this.cities);
-
-    this.changeDetector.detectChanges();
-  }
-
-  onCitySelected(selectedCityDto: SelectedCityDto | null) {
-    this.tripState.updateSelectedCity(selectedCityDto);
-    this.panelViewToSet = null;
-  }
-
-  onDayChanged(dayIndex: number): void {
-    this.tripState.updateCurrentDayIndex(dayIndex);
   }
 
   onViewChanged(viewData: { view: ModalView }): void {
@@ -193,11 +104,11 @@ export class CreateTripPageComponent implements OnInit, OnDestroy {
 
   confirmDelete() {
     if (this.selectedEntity?.type === 'city') {
-      this.deleteCity();
+      this.tripState.deleteCity(this.selectedEntity.data as AddCityDto);
     }
 
     if (this.selectedEntity?.type === 'waypoint') {
-      this.deleteWaypoint();
+      this.tripState.deleteWaypoint(this.selectedEntity.data as AddWaypointDto);
     }
 
     this.changeDetector.detectChanges();
@@ -206,116 +117,5 @@ export class CreateTripPageComponent implements OnInit, OnDestroy {
 
   closeDeleteModal() {
     this.selectedEntity = null;
-  }
-
-  private deleteCity() {
-    const cityIndex = this.cities.indexOf(
-      this.selectedEntity?.data as AddCityDto
-    );
-    if (cityIndex > -1) {
-      this.cities.splice(cityIndex, 1);
-    }
-
-    this.cities
-      .sort((x) => x.order)
-      .forEach((city, index) => {
-        city.order = index;
-
-        if (index == 0) {
-          city.arrivalDate = this.startDate;
-        } else {
-          let tempDate = new Date(this.cities[index - 1].arrivalDate!);
-          tempDate.setDate(
-            this.cities[index - 1].arrivalDate!.getDate() +
-              this.cities[index - 1].numberOfNights
-          );
-          city.arrivalDate = tempDate;
-        }
-      });
-    this.cities = [...this.cities];
-
-    console.log(
-      'City: ' + this.selectedEntity?.data.name + ' deleted succesfully.'
-    );
-  }
-
-  private deleteWaypoint() {
-    let city = this.cities.find(
-      (city) => city === this.selectedCity!.selectedCity
-    );
-
-    if (city === undefined) {
-      console.error('No city found.');
-      return;
-    }
-
-    const waypointIndex = city.waypoints[this.currentDayIndex].indexOf(
-      this.selectedEntity?.data as AddWaypointDto
-    );
-    if (waypointIndex > -1) {
-      city.waypoints[this.currentDayIndex].splice(waypointIndex, 1);
-    }
-
-    city.waypoints[this.currentDayIndex]
-      .sort((x) => x.order)
-      .forEach((waypoint, index) => {
-        waypoint.order = index;
-      });
-
-    this.cities = [...this.cities];
-    console.log(
-      'Waypoint: ' + this.selectedEntity?.data.name + ' deleted succesfully.'
-    );
-  }
-
-  private setWaypointsOrder(
-    waypoint: AddWaypointDto,
-    initialOrder: number,
-    city: AddCityDto
-  ) {
-    const waypointSplitStartTime = waypoint.startTime.split(':');
-    const waypointStartTime =
-      parseInt(waypointSplitStartTime[0], 10) * 60 +
-      parseInt(waypointSplitStartTime[1], 10);
-
-    let currentDayWaypoints = city.waypoints[this.currentDayIndex];
-
-    let order = initialOrder;
-    let index = 0;
-    let isLast = true;
-    while (index < currentDayWaypoints.length) {
-      const splitStartTime = currentDayWaypoints[index].startTime.split(':');
-      const currentWaypointStartTime =
-        parseInt(splitStartTime[0], 10) * 60 + parseInt(splitStartTime[1], 10);
-
-      if (waypointStartTime < currentWaypointStartTime) {
-        isLast = false;
-        break;
-      }
-
-      const currentPlaceId = currentDayWaypoints[index].placeId;
-      if (
-        waypointStartTime === currentWaypointStartTime &&
-        waypoint.placeId === currentPlaceId
-      ) {
-        break;
-      }
-
-      index++;
-    }
-
-    if (isLast === false) {
-      order = currentDayWaypoints[index].order;
-      for (
-        let secondIndex = index;
-        secondIndex < currentDayWaypoints.length;
-        secondIndex++
-      ) {
-        currentDayWaypoints[secondIndex].order =
-          currentDayWaypoints[secondIndex].order + 1;
-      }
-    }
-
-    waypoint.order = order;
   }
 }
